Collapse SerialiseLib into renamed StdLib
