+++ ulox/ulox.core.tests/Package/Scripts/Tests/StdLibrary.ulox
@@ -1,12 +1,5 @@
-build bind "CoreLibrary";
-
-test CoreLibraryTests
+test StdLibraryTests
 {
-    // testcase Print
-    // {
-    //     Assert.Fail();
-    // }
-
     testcase DuplicateMatches
     {
         var a = 1;
