+++ ulox/ulox.core/Package/Runtime/Engine/Engine.cs
@@ -11,8 +11,7 @@ namespace ULox
         {
             Context = executionContext;
             Context.VM.SetEngine(this);
-            Context.AddLibrary(new AssertLibrary());
-            Context.AddLibrary(new SerialiseLibrary());
+            Context.AddLibrary(new StdLibrary());
             Context.AddLibrary(new DebugLibrary());
             Context.AddLibrary(new DiLibrary());
             Context.AddLibrary(new FreezeLibrary());
@@ -41,7 +40,7 @@ namespace ULox
         {
             var context = new Context(new ScriptLocator(), new Program(), new Vm());
             var engine = new Engine(context);
-            engine.Context.AddLibrary(new CoreLibrary(x => context.Log(x)));
+            engine.Context.AddLibrary(new PrintLibrary(x => context.Log(x)));
             return engine;
         }
     }
